requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
nameparser==1.1.3
